refactor: update UpdateProjectUseCaseTest to use direct calls & the related imports
