+++ src/main/kotlin/di/UseCaseModule.kt
@@ -4,7 +4,7 @@ import logic.useCase.CreateTaskUseCase
 import org.example.logic.useCase.*
 import org.example.logic.useCase.DeleteProjectUseCase
 import org.example.logic.useCase.deleteTask.DeleteTaskUseCase
-import org.example.logic.useCase.updateProject.UpdateProjectUseCase
+import org.example.logic.useCase.UpdateProjectUseCase
 import org.example.logic.useCase.updateTask.UpdateTaskUseCase
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
