+++ src/main/kotlin/presentation/screens/ProjectsOverviewUI.kt
@@ -5,7 +5,7 @@ import org.example.logic.models.AuditLogEntityType
 import org.example.logic.models.Project
 import org.example.logic.useCase.*
 import org.example.logic.useCase.DeleteProjectUseCase
-import org.example.logic.useCase.updateProject.UpdateProjectUseCase
+import org.example.logic.useCase.UpdateProjectUseCase
 import org.example.presentation.role.ProjectScreensOptions
 import org.koin.java.KoinJavaComponent.getKoin
 import presentation.utils.TablePrinter
