+++ src/test/kotlin/logic/useCase/UpdateProjectUseCaseTest.kt
@@ -16,7 +16,7 @@ import org.example.logic.repositries.AuthenticationRepository
 import org.example.logic.repositries.ProjectRepository
 import org.example.logic.repositries.TaskRepository
 import org.example.logic.useCase.GetProjectTasksUseCase
-import org.example.logic.useCase.updateProject.UpdateProjectUseCase
+import org.example.logic.useCase.UpdateProjectUseCase
 import org.example.logic.utils.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -40,7 +40,7 @@ class UpdateProjectUseCaseTest {
         getProjectTasksUseCase = mockk(relaxed = true)
         taskRepository = mockk(relaxed = true)
         updateProjectUseCase = UpdateProjectUseCase(
-            projectRepository, auditLogRepository, authenticationRepository, getProjectTasksUseCase, taskRepository
+            projectRepository, auditLogRepository, authenticationRepository,
         )
     }
